split test_dsa_models_mtp into 4 files
